requirements.txt
astroid==1.4.5
colorama==0.3.7
IPy==0.83
lazy-object-proxy==1.2.2
py==1.4.31
pylint==1.5.5
pytest==3.0
PyYAML==3.11
requests==2.9.1
six==1.10.0
wrapt==1.10.8
python-dateutil==2.6.0
pytest-html==1.15.2

